完善管理员功能

+++ src/main/java/com/talresource/Talent_Recruitment/controller/ManagerController.java
@@ -1,12 +1,16 @@
 package com.talresource.Talent_Recruitment.controller;
 
+import com.talresource.Talent_Recruitment.service.CommentService;
 import com.talresource.Talent_Recruitment.service.NewsService;
 import com.talresource.Talent_Recruitment.service.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -19,35 +23,48 @@ public class ManagerController {
     @Autowired
     private PostService postService;
 
-    @RequestMapping("/managerOptions")
-    public String managerOptions(){
-        return "manager/managerOptions";
+    @Autowired
+    private CommentService commentService;
+
+    @RequestMapping("/managerCheck")
+    public String managerCheck(HttpSession session){
+        session.removeAttribute("key");
+        return "manager/managerCheck";
     }
 
     @RequestMapping("/manager1")
-    public String manager1(ModelMap map){
+    public String manager1(HttpSession session,ModelMap map){
+        if(session.getAttribute("key") == null)
+            return "redirect:managerCheck";
         map.addAttribute("NewsList", newsService.selectAllNews());
         return "manager/manager1";
     }
 
     @RequestMapping("/manager2")
-    public String manager2(ModelMap map){
+    public String manager2(HttpSession session, ModelMap map){
+        if(session.getAttribute("key") == null)
+            return "redirect:managerCheck";
         map.addAttribute("PostList", postService.selectAllPost());
         return "manager/manager2";
     }
 
     @RequestMapping("/deleteNews")
-    public String deleteNews(ModelMap map, String NewsID){
+    public String deleteNews(HttpSession session, ModelMap map, String NewsID){
+        if(session.getAttribute("key") == null)
+            return "redirect:managerCheck";
         int id = Integer.parseInt(NewsID);
         if (newsService.deleteNewsByID(id)){
             map.addAttribute("NewsList", newsService.selectAllNews());
             return "manager/manager1";
         }
-        return "manager/manager1";
+        return "error";
     }
 
     @RequestMapping("/saveNews")
-    public String saveNews(ModelMap map, String NewsID, String NewsTitle, String NewsContent){
+    public String saveNews(HttpSession session, ModelMap map, String NewsID, String NewsTitle, String NewsContent){
+
+        if(session.getAttribute("key") == null)
+            return "redirect:managerCheck";
 
         Date today = new Date();
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
@@ -73,14 +90,26 @@ public class ManagerController {
 
 
     @RequestMapping("/deletePost")
-    public String deletePost(ModelMap map, String PostID){
+    public String deletePost(HttpSession session, ModelMap map, String PostID){
+
+        if(session.getAttribute("key") == null)
+            return "redirect:managerCheck";
+
         int id = Integer.parseInt(PostID);
-        if (postService.deletePostByID(id)){
+        if (postService.deletePostByID(id) && commentService.deleteCommentByPost(id)){
             map.addAttribute("PostList", postService.selectAllPost());
             return "manager/manager2";
         }
-        map.addAttribute("PostList", postService.selectAllPost());
+        return "error";
+    }
 
-        return "manager/manager2";
+
+    @RequestMapping("/requestkey")
+    @ResponseBody
+    public boolean requestkey(HttpSession session){
+        System.out.println("1");
+        session.setAttribute("key", "true");
+
+        return true;
     }
 }
